feat: Reduced false positives for many warnings in claims.csv.
fix: Fixed a bug that caused plint to crash when checking titles in debug mode.
fix: More clearly labeled a message about restrictions that appears when there is only one independent claim.

fix: Added more -ing words that are not functional to avoid false positives.

fix: Fixed a bug in the check that dependent claims start with "The".